Finished unittests
