+++ adsft/tests/test_tasks.py
@@ -5,7 +5,6 @@ import json
 from mock import patch
 import unittest
 from adsft import app, tasks
-from adsft.models import Base
 
 
 class TestWorkers(unittest.TestCase):
@@ -14,31 +13,18 @@ class TestWorkers(unittest.TestCase):
         unittest.TestCase.setUp(self)
         self.proj_home = os.path.join(os.path.dirname(__file__), '../..')
         self._app = tasks.app
-        self.app = app.ADSFulltextCelery('test',
+        self.app = app.ADSFulltextCelery('test', local_config=\
             {
-            'SQLALCHEMY_URL': 'sqlite:///',
-            'SQLALCHEMY_ECHO': False
             })
         tasks.app = self.app # monkey-patch the app object
-        Base.metadata.bind = self.app._session.get_bind()
-        Base.metadata.create_all()
     
     
     def tearDown(self):
         unittest.TestCase.tearDown(self)
-        Base.metadata.drop_all()
         self.app.close_app()
         tasks.app = self._app
 
 
-    def test_task_hello_world(self):
-        
-        # just for illustration how to mock multiple objects in one go
-        with patch.object(self.app, 'close_app') as example_method, \
-            patch.object(tasks.logger, 'info') as logger:
-            
-            tasks.task_hello_world({'name': 'Elgar'})
-            self.assertTrue('Hello Elgar we have recorded' in logger.call_args[0][0])
