+++ config.py
@@ -7,107 +7,6 @@ OUTPUT_CELERY_BROKER = 'pyamqp://guest:guest@localhost:6672/master_pipeline'
 OUTPUT_TASKNAME = 'adsmp.tasks.task_update_record'
 
 
-FULLTEXT_EXTRACT_PATH = '/vagrant/live'
+FULLTEXT_EXTRACT_PATH = './live'
 
-# TODO(rca): turn into real constants
-CONSTANTS = {
-    'META_PATH': 'meta_path',
-    'FILE_SOURCE': 'ft_source',
-    'BIBCODE': 'bibcode',
-    'PROVIDER': 'provider',
-    'UPDATE': 'UPDATE',
-    'FULL_TEXT': 'fulltext',
-    'FORMAT': 'file_format',
-    'TIME_STAMP': 'index_date',
-    'ACKNOWLEDGEMENTS': 'acknowledgements',
-    'DATASET': 'dataset',
-}
 
-META_CONTENT = {
-    'xml': {
-        'fulltext': {
-            'xpath': ['//body',
-                      '//section[@type="body"]',
-                      '//journalarticle-body'
-                      ],
-            'type': 'string',
-            'info': '',
-            },
-        'acknowledgements': {
-            'xpath': ['//ack',
-                      '//section[@type="acknowledgments"]',
-                      '//subsection[@type="acknowledgement" '
-                      'or @type="acknowledgment"]'
-                      ],
-            'type': 'string',
-            'info': '',
-            },
-        'dataset': {
-            'xpath': ['//named-content[@content-type="dataset"]'],
-            'type': 'list',
-            'info': 'xlink:href',
-        }
-    },
-    'teixml': {
-        'fulltext': {
-            'xpath': ['//body',
-                      ],
-            'type': 'string',
-            'info': '',
-            },
-        'acknowledgements': {
-            'xpath': ['//div[@type="acknowledgement"]',
-                      ],
-            'type': 'string',
-            'info': '',
-            },
-    },
-    'xmlelsevier': {
-        'fulltext': {
-            'xpath': ['//body',
-                      '//raw-text',
-                      ],
-            'type': 'string',
-            'info': '',
-            },
-        'acknowledgements': {
-            'xpath': ['//acknowledgment',
-                      '//ack',
-                      '//section[@type="acknowledgments"]',
-                      '//subsection[@type="acknowledgement" '
-                      'or @type="acknowledgment"]',
-                      '//*[local-name()="acknowledgment"]'
-                      ],
-            'type': 'string',
-            'info': '',
-            },
-        'dataset': {
-            'xpath': ['//named-content[@content-type="dataset"]'],
-            'type': 'list',
-            'info': 'xlink:href',
-        }
-    },
-    'html': {
-        'introduction': [
-            '//h2[contains(.,"ntroduction")]',
-            '//h3[contains(.,"ntroduction")]',
-            '//p[contains(.,"Abstract")]',
-            ],
-        'references': [
-            '//h2[contains(.,"References")]'
-        ],
-        'table': [
-            '//table'
-        ],
-        'table_links': [
-            '//a[contains(@href, "TABLE_NAME")]'
-        ],
-        'head': [
-            '//head'
-        ]
-    },
-    'txt': {'fulltext': ['']},
-    'ocr': {'fulltext': ['']},
-    'http': {'fulltext': ['']},
-    'pdf': {'fulltext': ['']},
-}
