+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+addopts = --cov=adsft --cov-report=term-missing
+testpaths = ./adsft/tests
