+++ requirements.txt
@@ -3,6 +3,6 @@ ConcurrentLogHandler==0.9.1
 dateutils==0.6.6
 git+https://github.com/adsabs/ADSPipelineUtils.git@master
 httpretty==0.8.14
-lxml==3.8.0
+lxml==3.4.2
 ptree==0.1
 requests==2.18.1
